Add a simple dotjs plugin system
Ever since the dotjs project got abandoned I've been looking for
something similar. It works like dis:

Whenever a webpage is loaded we look for a file in `~/.js/<hostname>.js`
and if we find one we read it's contents and execute the javascript we
find there on the page we just loaded.

I think the quote on the original dotjs project describes the
functionality pretty well.

> "I almost wish you could just stick JavaScript in ~/.js. Do you know
what I'm saying?"

- https://github.com/defunkt/dotjs